// frontend/src/api/generated/schemas/predictionConfigInput.ts
// @ts-nocheck
// This file is auto-generated by orval. Do not edit manually.
import type { FactorConfigInput } from "./factorConfigInput";
import type { PredictionConfigInputDate } from "./predictionConfigInputDate";

/**
 * Request body for prediction preview with custom factor config.
 */
export interface PredictionConfigInput {
  activation?: FactorConfigInput;
  /** 分析日期 */
  date?: PredictionConfigInputDate;
  divergence?: FactorConfigInput;
  momentum?: FactorConfigInput;
  relative_strength?: FactorConfigInput;
  rsi?: FactorConfigInput;
}